+++ alinkjs/packages/assetlink/src/AssetLinkPluginLoaderCore.js
@@ -1,14 +1,12 @@
 import { markRaw, reactive } from 'vue';
 
-import { v4 as uuidv4 } from 'uuid';
-
 import { loadModule } from 'vue3-sfc-loader/dist/vue3-sfc-loader.esm.js';
 import { parseComponent } from 'vue-template-compiler';
 
 import VuePluginShorthandDecorator from '@/VuePluginShorthandDecorator';
 import { default as pluginModuleLibrary, pluginModuleLibraryNames } from '@/pluginModuleLibrary';
 
-import { createDrupalUrl, currentEpochSecond, EventBus } from "assetlink-plugin-api";
+import { createDrupalUrl, currentEpochSecond, EventBus, uuidv4 } from "assetlink-plugin-api";
 
 export default class AssetLinkPluginLoaderCore {